requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coverage==7.11.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.6
dateparser==1.1.8
decorator==5.2.1
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.12.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
g2pkk==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
hangul-romanize==0.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
HyperPyYAML==1.2.2
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.45.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
neo4j==6.0.2
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai-whisper==20250625
packaging==25.0
pandas==1.5.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.10
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.11.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
sqlparse==0.5.3
srsly==2.5.1
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.57.1
triton==2.1.0
TTS==0.22.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==2.0.0
yarl==1.22.0
